# gen/http/openapi3.json
{"openapi":"3.0.3","info":{"title":"Directory Service","description":"Service for multiplying numbers, a Goa teaser","version":"1.0"},"servers":[{"url":"http://localhost:8000"}],"paths":{"/ds":{"get":{"tags":["ds"],"summary":"list ds","description":"List all stored bottles","operationId":"ds#list","responses":{"200":{"description":"OK response.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AccountMgmtResponseCollection"},"example":[{"accountname":"ABC Pvt Ltd","clusterurl":"John Doe","id":1,"uuid":"kmpySmUISimoRrJL6NL73w"},{"accountname":"ABC Pvt Ltd","clusterurl":"John Doe","id":1,"uuid":"kmpySmUISimoRrJL6NL73w"},{"accountname":"ABC Pvt Ltd","clusterurl":"John Doe","id":1,"uuid":"kmpySmUISimoRrJL6NL73w"},{"accountname":"ABC Pvt Ltd","clusterurl":"John Doe","id":1,"uuid":"kmpySmUISimoRrJL6NL73w"}]}}}}}},"/ds/complete/{token}":{"put":{"tags":["ds"],"summary":"complete ds","description":"Complete New Signup flow with token","operationId":"ds#complete","parameters":[{"name":"token","in":"path","description":"Verification token","required":true,"schema":{"type":"string","description":"Verification token","example":"Molestiae sunt placeat blanditiis."},"example":"Est enim eius iure."}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CompleteRequestBody"},"example":{"gaClientId":"Quisquam et assumenda aut aperiam voluptas saepe.","referer":"Dolorem enim officia dolorum molestiae aut placeat.","visitorToken":"Consequatur tenetur voluptatibus accusamus aut."}}}},"responses":{"200":{"description":"OK response.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CompleteResponseBody"},"example":{"accountname":"Harness","clusterurl":"http://localdev.harness.io/prod1","email":"shashank.singh@harness.io","name":"Shashank Singh","uuid":"kmpySmUISimoRrJL6NL73w"}}}}}}},"/ds/multiply/{a}/{b}":{"get":{"tags":["ds"],"summary":"demo ds","operationId":"ds#demo","parameters":[{"name":"a","in":"path","description":"Left operand","required":true,"schema":{"type":"integer","description":"Left operand","example":4178647503179819338,"format":"int64"},"example":8580616938411114698},{"name":"b","in":"path","description":"Right operand","required":true,"schema":{"type":"integer","description":"Right operand","example":8218317442234526001,"format":"int64"},"example":2644045839809312112}],"responses":{"200":{"description":"OK response.","content":{"application/json":{"schema":{"type":"integer","example":3960059040878664830,"format":"int64"},"example":3057605669671577677}}}}}},"/ds/openapi.json":{"get":{"tags":["ds"],"summary":"Download ./gen/http/openapi.json","operationId":"ds#/ds/openapi.json","responses":{"200":{"description":"File downloaded"}}}}},"components":{"schemas":{"AccountMgmtResponse":{"type":"object","properties":{"accountname":{"type":"string","description":"Name of the Account/Company","example":"ABC Pvt Ltd"},"clusterurl":{"type":"string","description":"Tells in which cluster does this account operates in","example":"John Doe"},"id":{"type":"integer","description":"ID is an integer type unique id of this DB entity of account.","example":1,"format":"int64"},"uuid":{"type":"string","description":"UUID is the the unique id of this Account","example":"kmpySmUISimoRrJL6NL73w"}},"description":"AccountMgmt type describes a customer account of company. (default view)","example":{"accountname":"ABC Pvt Ltd","clusterurl":"John Doe","id":1,"uuid":"kmpySmUISimoRrJL6NL73w"},"required":["id","uuid","clusterurl","accountname"]},"AccountMgmtResponseCollection":{"type":"array","items":{"$ref":"#/components/schemas/AccountMgmtResponse"},"description":"ListResponseBody is the result type for an array of AccountMgmtResponse (default view)","example":[{"accountname":"ABC Pvt Ltd","clusterurl":"John Doe","id":1,"uuid":"kmpySmUISimoRrJL6NL73w"},{"accountname":"ABC Pvt Ltd","clusterurl":"John Doe","id":1,"uuid":"kmpySmUISimoRrJL6NL73w"},{"accountname":"ABC Pvt Ltd","clusterurl":"John Doe","id":1,"uuid":"kmpySmUISimoRrJL6NL73w"}]},"CompleteRequestBody":{"type":"object","properties":{"gaClientId":{"type":"string","description":"gaClientId","example":"Et sit est sed consequatur."},"referer":{"type":"string","description":"referer","example":"Facilis libero nulla aut dolores quis."},"visitorToken":{"type":"string","description":"visitorToken","example":"Dolores doloremque quia voluptatibus rerum voluptatibus."}},"example":{"gaClientId":"Ullam hic non.","referer":"In expedita.","visitorToken":"In animi voluptatem."}},"CompleteResponseBody":{"type":"object","properties":{"accountname":{"type":"string","description":"Name of the Account/Company","example":"Harness"},"clusterurl":{"type":"string","description":"Tells in which cluster does this account operates in","example":"http://localdev.harness.io/prod1"},"email":{"type":"string","description":"email of this user","example":"shashank.singh@harness.io"},"name":{"type":"string","description":"name of this user","example":"Shashank Singh"},"uuid":{"type":"string","description":"UUID is the the unique id of this Account","example":"kmpySmUISimoRrJL6NL73w"}},"description":"CompleteResponseBody result type (default view)","example":{"accountname":"Harness","clusterurl":"http://localdev.harness.io/prod1","email":"shashank.singh@harness.io","name":"Shashank Singh","uuid":"kmpySmUISimoRrJL6NL73w"},"required":["uuid","email","name","clusterurl","accountname"]}}},"tags":[{"name":"ds","description":"The directory service handles account mgmt"}]}